notification-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.25.2
python-multipart==0.0.6
aiosmtplib==3.0.1
jinja2==3.1.2
aiosqlite==0.19.0
greenlet==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
